chore: move s3 content to lib
