+++ requirements.in
@@ -12,3 +12,4 @@ requests
 jsonschema
 cryptography
 boto3
+boto3-stubs[s3]
