+++ requirements.txt
@@ -12,12 +12,16 @@ attrs==23.2.0
     # via
     #   jsonschema
     #   referencing
-boto3==1.34.83
+boto3==1.34.84
     # via -r requirements.in
-botocore==1.34.83
+boto3-stubs[s3]==1.34.84
+    # via -r requirements.in
+botocore==1.34.84
     # via
     #   boto3
     #   s3transfer
+botocore-stubs==1.34.69
+    # via boto3-stubs
 certifi==2024.2.2
     # via
     #   httpcore
@@ -35,7 +39,7 @@ cryptography==42.0.5
     # via -r requirements.in
 h11==0.14.0
     # via httpcore
-httpcore==1.0.4
+httpcore==1.0.5
     # via httpx
 httpx==0.27.0
     # via lightkube
@@ -66,24 +70,26 @@ lightkube-models==1.29.0.7
     #   lightkube
 markupsafe==2.1.5
     # via jinja2
+mypy-boto3-s3==1.34.65
+    # via boto3-stubs
 ops==2.12.0
     # via
     #   -r requirements.in
     #   cosl
     #   ops-scenario
-ops-scenario==6.0.2
+ops-scenario==6.0.3
     # via pytest-interface-tester
-packaging==23.2
+packaging==24.0
     # via pytest
 pluggy==1.4.0
     # via pytest
-pycparser==2.21
+pycparser==2.22
     # via cffi
-pydantic==2.6.4
+pydantic==2.7.0
     # via
     #   -r requirements.in
     #   pytest-interface-tester
-pydantic-core==2.16.3
+pydantic-core==2.18.1
     # via pydantic
 pyhcl==0.4.5
     # via -r requirements.in
@@ -99,7 +105,7 @@ pyyaml==6.0.1
     #   lightkube
     #   ops
     #   ops-scenario
-referencing==0.33.0
+referencing==0.34.0
     # via
     #   jsonschema
     #   jsonschema-specifications
@@ -111,7 +117,7 @@ rpds-py==0.18.0
     # via
     #   jsonschema
     #   referencing
-s3transfer==0.10.0
+s3transfer==0.10.1
     # via boto3
 six==1.16.0
     # via python-dateutil
@@ -121,12 +127,18 @@ sniffio==1.3.1
     #   httpx
 typer==0.7.0
     # via pytest-interface-tester
-typing-extensions==4.10.0
+types-awscrt==0.20.5
+    # via botocore-stubs
+types-s3transfer==0.10.0
+    # via boto3-stubs
+typing-extensions==4.11.0
     # via
+    #   boto3-stubs
     #   cosl
+    #   mypy-boto3-s3
     #   pydantic
     #   pydantic-core
-urllib3==2.0.7
+urllib3==2.2.1
     # via
     #   botocore
     #   requests
